+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2020-11-12
+
+### Added
+
+- Loading text for webview
+- Added a test for the extension
+
+### Fixed
+
+- For Linux, there was an issue with the file path for user extensions, which lead to the extension not loading.
+
 ## [0.5.0] - 2020-09-28
 
 ### Added